Fixed pNode and pView so it has a snazzier and cleaner pView appearance.  Looks cleaner and snazzier than the standard TreeView appearance.  Right now it does not handle some of the other more complex attributes of a TreeView such as disabled, checked, etc.., but planning on adding those soon.
Also working on completing the coding for the namespaces and attributes.  I want to be able to go down the rabit hole when I add a prefix to a certain node.  I need a locking function to be enabled so if I decide to change a namespace for a specific node a parent can't changed that lock and apply its own namespace or schema definition.

Also adding some different types of xml types to a specific node.  So if a node or master node is a xhtml file, it will have an attribute distinguishing itself as being xhtml.   Same goes for xquery, xslt, xsd,xml, etc... I want to be able to know what it is so I can apply a function relating to it, and potential combine it with other nodes.  The goal of pWord is to add plug-ins for pWord so it can combine multiple other nodes together to create a new node or node set.